Sperated happy-path from failure test
