+++ tests/release/release.go
@@ -9,7 +9,6 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/redhat-appstudio/e2e-tests/pkg/framework"
 	gitopsv1alpha1 "github.com/redhat-appstudio/managed-gitops/appstudio-shared/apis/appstudio.redhat.com/v1alpha1"
-	"github.com/redhat-appstudio/release-service/api/v1alpha1"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	"knative.dev/pkg/apis"
@@ -26,11 +25,8 @@ const (
 	releasePipelineBundle = "quay.io/hacbs-release/demo:m5-alpine"
 	releaseStrategyPolicy = "policy"
 
-	avgPipelineCompletionTime     = 10 * time.Minute
-	defaultInterval               = 100 * time.Millisecond
-	failure1ApplicationName       = "test"
-	failure1SourceReleaseLinkName = "test-releaselink"
-	failure1ReleaseName           = "test-release"
+	avgPipelineCompletionTime = 10 * time.Minute
+	defaultInterval           = 100 * time.Millisecond
 )
 
 var snapshotComponents = []gitopsv1alpha1.ApplicationSnapshotComponent{
@@ -39,7 +35,7 @@ var snapshotComponents = []gitopsv1alpha1.ApplicationSnapshotComponent{
 	{"component-3", "quay.io/redhat-appstudio/component3@sha256:d90a0a33e4c5a1daf5877f8dd989a570bfae4f94211a8143599245e503775b1f"},
 }
 
-var _ = framework.ReleaseSuiteDescribe("test-demo", func() {
+var _ = framework.ReleaseSuiteDescribe("test-release-service-happy-path", func() {
 	defer GinkgoRecover()
 	// Initialize the tests controllers
 	framework, err := framework.NewFramework()
@@ -47,9 +43,6 @@ var _ = framework.ReleaseSuiteDescribe("test-demo", func() {
 
 	var devNamespace = uuid.New().String()
 	var managedNamespace = uuid.New().String()
-
-	var failureDevNamespace = "user-" + uuid.New().String()
-	var failureManagedNamespace = "managed-" + uuid.New().String()
 
 	BeforeAll(func() {
 		// Create the dev namespace
@@ -142,72 +135,4 @@ var _ = framework.ReleaseSuiteDescribe("test-demo", func() {
 		})
 	})
 
-	var _ = Describe("Failure- Missing matching ReleaseLink", func() {
-		BeforeAll(func() {
-			// Create the dev namespace
-			sourceNamespace, err := framework.CommonController.CreateTestNamespace(failureDevNamespace)
-			Expect(err).NotTo(HaveOccurred(), "Error when creating namespace '%s': %v", sourceNamespace.Name, err)
-
-			// Create the managed namespace
-			managedNamespace, err := framework.CommonController.CreateTestNamespace(failureManagedNamespace)
-			Expect(err).NotTo(HaveOccurred(), "Error when creating namespace '%s': %v", managedNamespace.Name, err)
-		})
-
-		AfterAll(func() {
-			// Delete the dev and managed namespaces with all the resources created in them
-			Expect(framework.CommonController.DeleteNamespace(failureDevNamespace)).NotTo(HaveOccurred())
-			Expect(framework.CommonController.DeleteNamespace(failureManagedNamespace)).NotTo(HaveOccurred())
-		})
-
-		var _ = Describe("All required resources are created successfully", func() {
-			It("Create an ApplicationSnapshot for M5 application", func() {
-				_, err := framework.ReleaseController.CreateApplicationSnapshot(snapshotName, failureDevNamespace, failure1ApplicationName, snapshotComponents)
-				Expect(err).NotTo(HaveOccurred())
-			})
-
-			It("Create ReleaseLink in source namespace", func() {
-				_, err := framework.ReleaseController.CreateReleaseLink(failure1SourceReleaseLinkName, failureDevNamespace, failure1ApplicationName, failureManagedNamespace, "")
-				Expect(err).NotTo(HaveOccurred())
-			})
-
-			It("Create a Release in source namespace", func() {
-				_, err := framework.ReleaseController.CreateRelease(failure1ReleaseName, failureDevNamespace, snapshotName, failure1SourceReleaseLinkName)
-				Expect(err).NotTo(HaveOccurred())
-			})
-		})
-
-		var _ = Describe("A ReleaseLink has to have a matching one in a managed workspace", func() {
-			release := &v1alpha1.Release{}
-			releaseReason := ""
-			releaseMessage := ""
-
-			It("The Release have failed with the REASON field set to ReleaseValidationError", func() {
-				Eventually(func() bool {
-					release, err = framework.ReleaseController.GetRelease(failure1ReleaseName, failureDevNamespace)
-
-					if err != nil || release == nil {
-						return false
-					}
-
-					releaseReason = release.Status.Conditions[0].Reason
-
-					return release.IsDone() && meta.IsStatusConditionFalse(release.Status.Conditions, "Succeeded") && Expect(releaseReason).To(Equal("ReleaseValidationError"))
-				}, avgPipelineCompletionTime, defaultInterval).Should(BeTrue())
-			})
-
-			It("Condition message describes an error finding a matching ReleaseLink", func() {
-				Eventually(func() bool {
-					release, err = framework.ReleaseController.GetRelease(failure1ReleaseName, failureDevNamespace)
-
-					if err != nil || release == nil {
-						return false
-					}
-
-					releaseMessage = release.Status.Conditions[0].Message
-
-					return Expect(releaseMessage).Should(ContainSubstring("no ReleaseLink found in target workspace"))
-				}, avgPipelineCompletionTime, defaultInterval).Should(BeTrue())
-			})
-		})
-	})
 })
